Rework suggestions dropdown to show "No matches!" when there aren't any.
The need to differentiate between "no matches because no search string"
versus "no matches because search string returns null" meant we need
to put a copy of the current search string in the model.

And since we introduced that, the other model data got moved into the
component proper, too.  Which should help with reuse.